Build trait cathetral
