+++ reed-solomon-novelpoly/src/field/inc_log_mul.rs
@@ -1,46 +1,47 @@
-use derive_more::{Add, AddAssign, BitXor, BitXorAssign, Sub, SubAssign};
 
-/// Additive via XOR form of f2e16
-#[derive(Clone, Copy, Debug, Default, BitXor, BitXorAssign, PartialEq, Eq)] // PartialOrd,Ord
-pub struct Additive(pub Elt);
+pub use core::ops::{Mul, MulAssign};
 
-impl Additive {
-    #[inline(always)]
-	pub fn to_wide(self) -> Wide {
-		self.0 as Wide
-	}
-    #[inline(always)]
-	pub fn from_wide(x: Wide) -> Additive {
-		Additive(x as Elt)
-	}
+include!("inc_cantor_basis.rs");
 
-	pub const ZERO: Additive = Additive(0);
-}
-
-#[cfg(table_bootstrap_complete)]
-impl Additive {
-	/// Return multiplier prepared form
-    #[inline(always)]
-	pub fn to_multiplier(self) -> Multiplier {
-		Multiplier(LOG_TABLE[self.0 as usize])
-	}
+impl Mul<Multiplier> for Additive {
+    type Output = Additive;
 
 	/// Return a*EXP_TABLE[b] over GF(2^r)
     #[inline(always)]
-	pub fn mul(self, other: Multiplier) -> Additive {
+    #[cfg(table_bootstrap_complete)]
+    fn mul(self, other: Multiplier) -> Additive {
 		if self == Self::ZERO {
 			return Self::ZERO;
 		}
 		let log = (LOG_TABLE[self.0 as usize] as Wide) + other.0 as Wide;
 		let offset = (log & ONEMASK as Wide) + (log >> FIELD_BITS);
 		Additive(EXP_TABLE[offset as usize])
+    }
+
+    #[cfg(not(table_bootstrap_complete))]
+    fn mul(self, other: Multiplier) -> Additive { panic!(); }
+}
+
+impl MulAssign<Multiplier> for Additive {
+    #[inline(always)]
+    fn mul_assign(&mut self, rhs: Multiplier) {
+        *self = (*self) * rhs;
+    }
+}
+
+#[cfg(table_bootstrap_complete)]
+impl FieldMul<Multiplier> for Additive {
+	/// Return multiplier prepared form
+    #[inline(always)]
+	fn to_multiplier(self) -> Multiplier {
+		Multiplier(LOG_TABLE[self.0 as usize])
 	}
 
 	/// Multiply field elements by a single multiplier, using SIMD if available
     #[inline(always)]
-	pub fn mul_assign_slice(selfy: &mut [Self], other: Multiplier) {
+	fn mul_assign_slice(selfy: &mut [Self], other: Multiplier) {
 		for s in selfy {
-			*s = s.mul(other);
+			*s *= other;
 		}
 	}
 }
@@ -61,78 +62,6 @@ impl Multiplier {
 	}
 }
 
-
-/// Fast Walsh–Hadamard transform over modulo ONEMASK
-pub fn walsh(data: &mut [Multiplier], size: usize) {
-	let mut depart_no = 1_usize;
-	while depart_no < size {
-		let mut j = 0;
-		let depart_no_next = depart_no << 1;
-		while j < size {
-			for i in j..(depart_no + j) {
-				// We deal with data in log form here, but field form looks like:
-				//			 data[i] := data[i] / data[i+depart_no]
-				// data[i+depart_no] := data[i] * data[i+depart_no]
-				let mask = ONEMASK as Wide;
-				let tmp2: Wide = data[i].to_wide() + mask - data[i + depart_no].to_wide();
-				let tmp1: Wide = data[i].to_wide() + data[i + depart_no].to_wide();
-				data[i] = Multiplier(((tmp1 & mask) + (tmp1 >> FIELD_BITS)) as Elt);
-				data[i + depart_no] = Multiplier(((tmp2 & mask) + (tmp2 >> FIELD_BITS)) as Elt);
-			}
-			j += depart_no_next;
-		}
-		depart_no = depart_no_next;
-	}
-}
-
-#[allow(unused)]
-fn bitpoly_mul(mut a: Wide, mut b: Wide) -> Wide {
-    let mut r: Wide =0;
-    for i in 0..FIELD_BITS {
-        if (b>>i) & 1 != 0 {
-			r ^= a<<i;
-		}
-    }
-    r
-}
-
-#[allow(unused)]
-fn gf_mul_bitpoly_reduced(a: Elt, b: Elt) -> Elt {
-    use core::convert::TryInto;
-    let len = FIELD_BITS;
-    let mut r: Wide = bitpoly_mul(a as Wide,b as Wide);
-    let red : Wide = (1 << FIELD_BITS) + (GENERATOR as Wide);
-    for i in (len..=(len*2-1)).rev() {
-        if r & (1<<i) != 0 {
-			r ^= red<<(i-len);
-		}
-    }
-    r.try_into().unwrap()
-}
-
-#[test]
-fn is_cantor_basis() {
-    for w in BASE.windows(2) {
-        let b = w[1];
-        let square = gf_mul_bitpoly_reduced(b,b);
-        let a = w[0];
-        // let eq = if a == (square ^ b) { "==" } else { "!=" };
-        // println!("{:#b} {} {:#b}\n", a, eq, square ^ b);
-        assert_eq!(a, square ^ b);
-    }
-}
-
-fn generate_cantor_basis(mut next: Elt) -> Option<[Elt; FIELD_BITS]> {
-    let mut base: [Elt; FIELD_BITS] = [0; FIELD_BITS];
-    base[0] = 1;
-    for b in base.iter_mut().rev() {
-        if next == 0 || (next == 1 && *b != 1) { return None; }
-        *b = next;
-        let square = gf_mul_bitpoly_reduced(next,next);
-        next ^= square;
-    }
-    if next == 0 { Some(base) } else { None }
-}
 
 /*
 Actually our to_multiplier abstraction is leaky
@@ -166,9 +95,9 @@ fn embedded_gf16() {
 #[test]
 fn print_gf256() {
     use std::io::Write;
-	let mut w = std::fs::OpenOptions::new().create(true).truncate(true).write(true).open(FIELD_NAME).unwrap();
+	let mut w = std::fs::OpenOptions::new().create(true).truncate(true).write(true).open(Additive::FIELD_NAME).unwrap();
 
-    write!(w, "\n\n\n{} :\n", FIELD_NAME);
+    write!(w, "\n\n\n{} :\n", Additive::FIELD_NAME);
     for i in 1..=255 {
         write!(w, "{:#b} * .. = ", i);
         let i = Additive(i).to_multiplier();
