+++ reed-solomon-novelpoly/src/field/mod.rs
@@ -1,14 +1,20 @@
+
 use crate::errors::*;
 use crate::util::*;
 
 #[macro_use]
-mod gen;
+mod traits;
+pub use traits::{FieldAdd,FieldMul};
+
 pub mod f256;
 pub mod f2e16;
+pub mod afft;
 
 
 #[test]
 fn agreement_f2e16_with_f256() {
+    use core::ops::Mul;
+    
     for i in 1..=255 {
         let i_f256 = f256::Additive(i).to_multiplier();
         let i_f2e16 = f2e16::Additive(i as u16).to_multiplier();
